update gam module to use submodules pattern
