+++ modules/.submodules.json
@@ -4,6 +4,7 @@
     "id5IdSystem"
   ],
   "adpod": [
-    "freeWheelAdserverVideo"
+    "freeWheelAdserverVideo",
+    "dfpAdServerVideo"
   ]
 }
