+++ modules/adpod.js
@@ -22,6 +22,7 @@ import { config } from '../src/config';
 import { ADPOD } from '../src/mediaTypes';
 import Set from 'core-js/library/fn/set';
 import find from 'core-js/library/fn/array/find';
+import { auctionManager } from '../src/auctionManager';
 
 const from = require('core-js/library/fn/array/from');
 
@@ -393,6 +394,7 @@ function initAdpodHooks() {
 }
 
 initAdpodHooks()
+
 /**
  *
  * @param {Array[Object]} bids list of 'winning' bids that need to be cached
@@ -431,11 +433,142 @@ export function sortByPricePerSecond(a, b) {
   return 0;
 }
 
+/**
+ * This function returns targeting keyvalue pairs for freewheel adserver module
+ * @param {Object} options
+ * @param {Array[string]} codes
+ * @param {function} callback
+ * @returns targeting kvs for adUnitCodes
+ */
+export function getTargeting({codes, callback} = {}) {
+  if (!callback) {
+    utils.logError('No callback function was defined in the getTargeting call.  Aborting getTargeting().');
+    return;
+  }
+  codes = codes || [];
+  const adPodAdUnits = getAdPodAdUnits(codes);
+  const bidsReceived = auctionManager.getBidsReceived();
+  const competiveExclusionEnabled = config.getConfig('adpod.brandCategoryExclusion');
+  const deferCachingSetting = config.getConfig('adpod.deferCaching');
+  const deferCachingEnabled = (typeof deferCachingSetting === 'boolean') ? deferCachingSetting : true;
+
+  let bids = getBidsForAdpod(bidsReceived, adPodAdUnits);
+  bids = (competiveExclusionEnabled || deferCachingEnabled) ? getExclusiveBids(bids) : bids;
+  bids.sort(sortByPricePerSecond);
+
+  let targeting = {};
+  if (deferCachingEnabled === false) {
+    adPodAdUnits.forEach((adUnit) => {
+      let adPodTargeting = [];
+      let adPodDurationSeconds = utils.deepAccess(adUnit, 'mediaTypes.video.adPodDurationSec');
+
+      bids
+        .filter((bid) => bid.adUnitCode === adUnit.code)
+        .forEach((bid, index, arr) => {
+          if (bid.video.durationBucket <= adPodDurationSeconds) {
+            adPodTargeting.push({
+              [TARGETING_KEY_PB_CAT_DUR]: bid.adserverTargeting[TARGETING_KEY_PB_CAT_DUR]
+            });
+            adPodDurationSeconds -= bid.video.durationBucket;
+          }
+          if (index === arr.length - 1 && adPodTargeting.length > 0) {
+            adPodTargeting.push({
+              [TARGETING_KEY_CACHE_ID]: bid.adserverTargeting[TARGETING_KEY_CACHE_ID]
+            });
+          }
+        });
+      targeting[adUnit.code] = adPodTargeting;
+    });
+
+    callback(null, targeting);
+  } else {
+    let bidsToCache = [];
+    adPodAdUnits.forEach((adUnit) => {
+      let adPodDurationSeconds = utils.deepAccess(adUnit, 'mediaTypes.video.adPodDurationSec');
+
+      bids
+        .filter((bid) => bid.adUnitCode === adUnit.code)
+        .forEach((bid) => {
+          if (bid.video.durationBucket <= adPodDurationSeconds) {
+            bidsToCache.push(bid);
+            adPodDurationSeconds -= bid.video.durationBucket;
+          }
+        });
+    });
+
+    callPrebidCacheAfterAuction(bidsToCache, function(error, bidsSuccessfullyCached) {
+      if (error) {
+        callback(error, null);
+      } else {
+        let groupedBids = utils.groupBy(bidsSuccessfullyCached, 'adUnitCode');
+        Object.keys(groupedBids).forEach((adUnitCode) => {
+          let adPodTargeting = [];
+
+          groupedBids[adUnitCode].forEach((bid, index, arr) => {
+            adPodTargeting.push({
+              [TARGETING_KEY_PB_CAT_DUR]: bid.adserverTargeting[TARGETING_KEY_PB_CAT_DUR]
+            });
+
+            if (index === arr.length - 1 && adPodTargeting.length > 0) {
+              adPodTargeting.push({
+                [TARGETING_KEY_CACHE_ID]: bid.adserverTargeting[TARGETING_KEY_CACHE_ID]
+              });
+            }
+          });
+          targeting[adUnitCode] = adPodTargeting;
+        });
+
+        callback(null, targeting);
+      }
+    });
+  }
+  return targeting;
+}
+
+/**
+ * This function returns the adunit of mediaType adpod
+ * @param {Array} codes adUnitCodes
+ * @returns {Array[Object]} adunits of mediaType adpod
+ */
+function getAdPodAdUnits(codes) {
+  return auctionManager.getAdUnits()
+    .filter((adUnit) => utils.deepAccess(adUnit, 'mediaTypes.video.context') === ADPOD)
+    .filter((adUnit) => (codes.length > 0) ? codes.indexOf(adUnit.code) != -1 : true);
+}
+
+/**
+ * This function removes bids of same freewheel category. It will be used when competitive exclusion is enabled.
+ * @param {Array[Object]} bidsReceived
+ * @returns {Array[Object]} unique freewheel category bids
+ */
+function getExclusiveBids(bidsReceived) {
+  let bids = bidsReceived
+    .map((bid) => Object.assign({}, bid, {[TARGETING_KEY_PB_CAT_DUR]: bid.adserverTargeting[TARGETING_KEY_PB_CAT_DUR]}));
+  bids = utils.groupBy(bids, TARGETING_KEY_PB_CAT_DUR);
+  let filteredBids = [];
+  Object.keys(bids).forEach((targetingKey) => {
+    bids[targetingKey].sort(utils.compareOn('responseTimestamp'));
+    filteredBids.push(bids[targetingKey][0]);
+  });
+  return filteredBids;
+}
+
+/**
+ * This function returns bids for adpod adunits
+ * @param {Array[Object]} bidsReceived
+ * @param {Array[Object]} adPodAdUnits
+ * @returns {Array[Object]} bids of mediaType adpod
+ */
+function getBidsForAdpod(bidsReceived, adPodAdUnits) {
+  let adUnitCodes = adPodAdUnits.map((adUnit) => adUnit.code);
+  return bidsReceived
+    .filter((bid) => adUnitCodes.indexOf(bid.adUnitCode) != -1 && (bid.video && bid.video.context === ADPOD))
+}
+
 const sharedMethods = {
   TARGETING_KEY_PB_CAT_DUR: TARGETING_KEY_PB_CAT_DUR,
   TARGETING_KEY_CACHE_ID: TARGETING_KEY_CACHE_ID,
-  'sortByPricePerSecond': sortByPricePerSecond,
-  'callPrebidCacheAfterAuction': callPrebidCacheAfterAuction
+  'getTargeting': getTargeting
 }
 Object.freeze(sharedMethods);
